Naive attempt at updating `requests` library.
This project is currently vulnerable to CVE-2018-18074, through the dependency on the `requests` library. Upgrading to a version above 2.20 should address the issue. This is a naive attempt at upgrading straight to the latest version, 2.21.0.

I'm not a Pythonista, so I'm not sure if this will work. Worth a shot though.