Document how to upgrade a package
I'm not much of a Pythonista, so whenever I need to update a dependency
for a Python project as part of Puffin Security Guardian duty, I always
have to spend a small amount of time remembering how to do it. This
should make it easier the next time this occurs.